FEAT: Redirect user to where she was after verification
