+++ lib/decidim/verifications/csv_email/engine.rb
@@ -15,6 +15,14 @@ module Decidim
           root to: "authorizations#new"
         end
 
+        initializer "decidim_csv_email.add_authorization_handlers" do |_app|
+          Decidim::Verifications.register_workflow(:csv_email_authorization_handler) do |workflow|
+            workflow.form = "CsvEmailAuthorizationHandler"
+            # workflow.engine = Decidim::Verifications::CsvEmail::Engine
+            # workflow.admin_engine = Decidim::Verifications::CsvEmail::AdminEngine
+          end
+        end
+
         config.to_prepare do
           Dir.glob(Decidim::Verifications::CsvEmail::Engine.root + "app/decorators/**/*_decorator*.rb").each do |c|
             require_dependency(c)
