+++ lib/decidim/verifications/csv_email/version.rb
@@ -5,7 +5,7 @@ module Decidim
   module Verifications
     module CsvEmail
       def self.version
-        "0.0.11"
+        "0.1.0"
       end
 
       def self.decidim_version
